requirements.txt
boto3~=1.34.2
flask~=3.0.0
botocore~=1.34.2
PyJWT~=2.8.0